Sliding Tiles 3x3 (Greedy, A*)
Given a random placement of the sliding tiles in an 8-puzzle, solve the problem of rearranging them in a row-major order. The sliding tiles are numbered from 1 to 8 on a 3x3 board where one of the squares is blank. You are permitted to slide the tiles either horizontally or vertically into the blank square. Use greedy search and A* search as informed (heuristic) search strategies to find a solution.