Sudoku 9x9 (CSP)
Solve the Sudoku problem by using Constraint Satisfaction Problem (CSP). The goal of Sudoku is to fill a 9x9 board with digits so that each column, each row, and each of the nine 3x3 subgrids contains all of the digits from 1 to 9. The puzzle starts with some cells already filled in with numbers, and the objective is to complete the grid.